qa_profile/tests/requirements.txt

robotframework==3.2.1
requests